修正資料夾結構
